+++ packages/eas-cli/src/commands/channel/edit.ts
@@ -4,6 +4,7 @@ import { print } from 'graphql';
 import gql from 'graphql-tag';
 
 import { selectBranchOnAppAsync } from '../../branch/queries';
+import { hasStandardBranchMap } from '../../channel/branch-mapping';
 import { selectChannelOnAppAsync } from '../../channel/queries';
 import EasCommand from '../../commandUtils/EasCommand';
 import { ExpoGraphqlClient } from '../../commandUtils/context/contextUtils/createGraphqlClient';
@@ -18,6 +19,7 @@ import { BranchQuery } from '../../graphql/queries/BranchQuery';
 import { ChannelQuery } from '../../graphql/queries/ChannelQuery';
 import { UpdateChannelBasicInfoFragmentNode } from '../../graphql/types/UpdateChannelBasicInfo';
 import Log from '../../log';
+import { isRollout } from '../../rollout/branch-mapping';
 import { enableJsonOutput, printJsonOnlyOutput } from '../../utils/json';
 
 export async function updateChannelBranchMappingAsync(
@@ -98,8 +100,10 @@ export default class ChannelEdit extends EasCommand {
           paginatedQueryOptions: { json, nonInteractive, offset: 0 },
         });
 
-    if (existingChannel.updateBranches.length > 1) {
+    if (isRollout(existingChannel)) {
       throw new Error('There is a rollout in progress. Manage it with "channel:rollout" instead.');
+    } else if (!hasStandardBranchMap(existingChannel)) {
+      throw new Error('Only standard branch mappings can be edited with this command.');
     }
 
     const branch = branchFlag